mining_server/requirements.txt
pandas
nltk
fastapi
starlette
beautifulsoup4
requests
python-dotenv
uvicorn
httpx